+++ vite-plugin-ssr/node/helpers/addSsrMiddleware.ts
@@ -1,6 +1,6 @@
 export { addSsrMiddleware }
 
-import { renderPage } from '../../runtime/renderPage'
+import { renderPage } from '../runtime/renderPage'
 import type { ViteDevServer } from 'vite'
 
 type ConnectServer = ViteDevServer['middlewares']
